+++ src/app/services/osc.service.ts
@@ -2,13 +2,34 @@ import { Injectable } from '@angular/core';
 import { invoke } from '@tauri-apps/api';
 import { SleepService } from './sleep.service';
 import { OscScript, OscScriptSleepAction } from '../models/osc-script';
-import { cloneDeep } from 'lodash';
+import { cloneDeep, isEqual, pick } from 'lodash';
 import { TaskQueue } from '../utils/task-queue';
 import { debug, info } from 'tauri-plugin-log-api';
 import { listen } from '@tauri-apps/api/event';
 import { OSCMessage, OSCMessageRaw, parseOSCMessage } from '../models/osc-message';
-import { BehaviorSubject, filter, firstValueFrom, map, Observable, Subject, take, tap } from 'rxjs';
+import {
+  BehaviorSubject,
+  distinctUntilChanged,
+  filter,
+  firstValueFrom,
+  map,
+  Observable,
+  skip,
+  Subject,
+  take,
+  tap,
+} from 'rxjs';
 import { AppSettingsService } from './app-settings.service';
+import { AppSettings } from '../models/settings';
+import { isValidHostname, isValidIPv4, isValidIPv6 } from '../utils/regex-utils';
+
+export type OscAddressError = 'PORT_IN_USE' | 'PORT_INVALID' | 'PORT_IO_IDENTICAL' | 'INVALID_HOST';
+export type OscAddressValidation = {
+  oscReceivingHost?: OscAddressError[];
+  oscReceivingPort?: OscAddressError[];
+  oscSendingHost?: OscAddressError[];
+  oscSendingPort?: OscAddressError[];
+};
 
 @Injectable({
   providedIn: 'root',
@@ -17,6 +38,16 @@ export class OscService {
   private scriptQueue: TaskQueue = new TaskQueue({ runUniqueTasksConcurrently: true });
   private _messages: Subject<OSCMessage> = new Subject<OSCMessage>();
   public messages: Observable<OSCMessage> = this._messages.asObservable();
+  private validationLock: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
+  private _addressValidation: BehaviorSubject<OscAddressValidation> = new BehaviorSubject({});
+  public addressValidation: Observable<OscAddressValidation> =
+    this._addressValidation.asObservable();
+
+  private receivingFeaturesEnabled: Observable<boolean> = this.appSettings.settings.pipe(
+    map((settings) =>
+      [settings.oscEnableExpressionMenu, settings.oscEnableExternalControl].some(Boolean)
+    )
+  );
 
   private _initializedOnAddress: BehaviorSubject<string | null> = new BehaviorSubject<
     string | null
@@ -30,27 +61,141 @@ export class OscService {
     });
     this.appSettings.settings
       .pipe(
-        map(
-          (settings) => [settings.oscReceivingHost, settings.oscReceivingPort] as [string, number]
-        ),
-        take(1),
-        filter(([, port]) => port > 0 && port <= 65535),
-        tap(([host, port]) => this.init_receiver(host, port))
+        distinctUntilChanged((a, b) => {
+          const pickRelevant = (c: AppSettings) =>
+            pick(c, [
+              'oscEnableExpressionMenu',
+              'oscEnableExternalControl',
+              'oscReceivingHost',
+              'oscReceivingPort',
+              'oscSendingHost',
+              'oscSendingPort',
+            ]);
+          return isEqual(pickRelevant(a), pickRelevant(b));
+        })
       )
-      .subscribe();
+      .subscribe((settings) => {
+        this.setOscSendingAddress(settings.oscSendingHost, settings.oscSendingPort, settings);
+        this.setOscReceivingAddress(settings.oscReceivingHost, settings.oscReceivingPort, settings);
+      });
   }
 
-  async init_receiver(host: string, port: number): Promise<boolean> {
-    const receiveAddr = `${host}:${port}`;
-    if (this._initializedOnAddress.value === receiveAddr) return true;
-    const result = await invoke<boolean>('osc_init', { receiveAddr });
-    if (!result) {
-      info(`[OSC] Could not bind a UDP socket on ${receiveAddr}.`);
-      this._initializedOnAddress.next(null);
+  public async setOscReceivingAddress(
+    host: string,
+    port: number,
+    settings?: AppSettings
+  ): Promise<OscAddressValidation> {
+    // Await the validation lock to be released
+    await firstValueFrom(this.validationLock.pipe(filter((v) => !v)));
+    // Lock the validation lock
+    this.validationLock.next(true);
+    // Define validation
+    const validation: OscAddressValidation = cloneDeep(this._addressValidation.value);
+    validation.oscReceivingHost = [];
+    validation.oscReceivingPort = [];
+    validation.oscSendingPort = (validation.oscSendingPort ?? []).filter(
+      (e) => e !== 'PORT_IO_IDENTICAL'
+    );
+    // Get the current settings if not provided
+    if (!settings) settings = await firstValueFrom(this.appSettings.settings);
+    // Validate sending and receiving ports not being identical
+    if (settings.oscSendingHost === host && settings.oscSendingPort === port) {
+      validation.oscReceivingPort.push('PORT_IO_IDENTICAL');
+      validation.oscSendingPort.push('PORT_IO_IDENTICAL');
+    }
+    // Validate port being in range
+    if (port <= 0 || port > 65535) {
+      validation.oscReceivingPort.push('PORT_INVALID');
+    }
+    // Validate host
+    if (host === '' || !(isValidIPv6(host) || isValidIPv4(host) || isValidHostname(host))) {
+      validation.oscReceivingHost.push('INVALID_HOST');
+    }
+    // Validate port not already being bound (only if any receiving features are enabled)
+    const receivingFeaturesEnabled = await firstValueFrom(this.receivingFeaturesEnabled);
+    if (
+      receivingFeaturesEnabled &&
+      !validation.oscReceivingHost.length &&
+      !validation.oscReceivingPort.length
+    ) {
+      const result = await this.start_osc_server(host, port);
+      if (!result) {
+        validation.oscReceivingPort.push('PORT_IN_USE');
+      }
     } else {
-      this._initializedOnAddress.next(receiveAddr);
+      await this.stop_osc_server();
+    }
+    // Update the current settings with the new data
+    await this.appSettings.updateSettings({
+      oscReceivingHost: host,
+      oscReceivingPort: port,
+    });
+    // Return validation results
+    this._addressValidation.next(validation);
+    this.validationLock.next(false);
+    return validation;
+  }
+
+  public async setOscSendingAddress(
+    host: string,
+    port: number,
+    settings?: AppSettings
+  ): Promise<OscAddressValidation> {
+    // Await the validation lock to be released
+    await firstValueFrom(this.validationLock.pipe(filter((v) => !v)));
+    // Lock the validation lock
+    this.validationLock.next(true);
+    // Define validation
+    const validation: OscAddressValidation = cloneDeep(this._addressValidation.value);
+    validation.oscSendingHost = [];
+    validation.oscSendingPort = [];
+    validation.oscReceivingPort = (validation.oscSendingPort ?? []).filter(
+      (e) => e !== 'PORT_IO_IDENTICAL'
+    );
+    // Get the current settings if not provided
+    if (!settings) settings = await firstValueFrom(this.appSettings.settings);
+    // Validate sending and receiving ports not being identical
+    if (settings.oscReceivingHost === host && settings.oscReceivingPort === port) {
+      validation.oscReceivingPort.push('PORT_IO_IDENTICAL');
+      validation.oscSendingPort.push('PORT_IO_IDENTICAL');
+    }
+    // Validate port being in range
+    if (port <= 0 || port > 65535) {
+      validation.oscSendingPort.push('PORT_INVALID');
+    }
+    // Validate host
+    if (host === '' || !(isValidIPv6(host) || isValidIPv4(host) || isValidHostname(host))) {
+      validation.oscSendingHost.push('INVALID_HOST');
+    }
+    // Update the current settings with the new data
+    await this.appSettings.updateSettings({
+      oscSendingHost: host,
+      oscSendingPort: port,
+    });
+    // Return validation results
+    this._addressValidation.next(validation);
+    this.validationLock.next(false);
+    return validation;
+  }
+
+  private async start_osc_server(host: string, port: number): Promise<boolean> {
+    const receiveAddr = `${host}:${port}`;
+    let result = true;
+    if (this._initializedOnAddress.value !== receiveAddr) {
+      result = await invoke<boolean>('start_osc_server', { receiveAddr });
+      if (!result) {
+        info(`[OSC] Could not bind a UDP socket on ${receiveAddr}.`);
+        this._initializedOnAddress.next(null);
+      } else {
+        this._initializedOnAddress.next(receiveAddr);
+      }
     }
     return result;
+  }
+
+  private async stop_osc_server(): Promise<void> {
+    await invoke<boolean>('stop_osc_server');
+    this._initializedOnAddress.next(null);
   }
 
   async send_float(address: string, value: number) {