+++ src/app/views/dashboard-view/views/settings-view/settings-vrchat-tab/settings-vrchat-tab.component.ts
@@ -3,22 +3,10 @@ import { SettingsTabComponent } from '../settings-tab/settings-tab.component';
 import { AppSettingsService } from '../../../../../services/app-settings.service';
 import { hshrink, vshrink } from '../../../../../utils/animations';
 import { VRChatService, VRChatServiceStatus } from '../../../../../services/vrchat.service';
-import {
-  combineLatest,
-  debounceTime,
-  distinctUntilChanged,
-  map,
-  of,
-  startWith,
-  Subject,
-  switchMap,
-  takeUntil,
-  tap,
-} from 'rxjs';
+import { debounceTime, distinctUntilChanged, Subject, takeUntil, tap } from 'rxjs';
 import { CurrentUser as VRChatUser } from 'vrchat/dist';
 import { ModalService } from 'src/app/services/modal.service';
-import { OscService } from '../../../../../services/osc.service';
-import { isValidHostname, isValidIPv4, isValidIPv6 } from '../../../../../utils/regex-utils';
+import { OscAddressValidation, OscService } from '../../../../../services/osc.service';
 import { APP_SETTINGS_DEFAULT } from '../../../../../models/settings';
 
 @Component({
@@ -53,6 +41,10 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
   protected oscReceivingPortStatus: 'INIT' | 'OK' | 'CHECKING' | 'ERROR' = 'INIT';
   protected oscReceivingPortError?: string;
 
+  protected get someOSCFeaturesEnabled() {
+    return this.appSettings.oscEnableExpressionMenu || this.appSettings.oscEnableExternalControl;
+  }
+
   constructor(
     settingsService: AppSettingsService,
     private vrchat: VRChatService,
@@ -74,6 +66,13 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
     this.listenForSendingHostChanges();
     this.listenForSendingPortChanges();
     this.listenForSettingsChanges();
+    this.listenForValidationChanges();
+  }
+
+  listenForValidationChanges() {
+    this.osc.addressValidation
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((validation) => this.processValidation(validation));
   }
 
   listenForSettingsChanges() {
@@ -102,7 +101,6 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
       .pipe(
         takeUntil(this.destroy$),
         distinctUntilChanged(),
-        switchMap((value) => combineLatest([of(value)]).pipe(map(([value]) => value))),
         tap(() => {
           this.oscReceivingHostStatus = 'CHECKING';
           this.oscReceivingHostError = undefined;
@@ -110,24 +108,8 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
         debounceTime(300)
       )
       .subscribe(async (host) => {
-        // Validate host
-        if (host === '' || !(isValidIPv6(host) || isValidIPv4(host) || isValidHostname(host))) {
-          this.oscReceivingHostStatus = 'ERROR';
-          this.oscReceivingHostError = 'invalidHost';
-          return;
-        }
-        // Try to bind
-        if (!(await this.osc.init_receiver(host, this.oscReceivingPort))) {
-          this.oscReceivingPortStatus = 'ERROR';
-          this.oscReceivingPortError = 'bindFailed';
-          return;
-        }
-        // Save new host
         this.oscReceivingHost = host;
-        this.oscReceivingHostStatus = 'OK';
-        this.settingsService.updateSettings({
-          oscReceivingHost: host,
-        });
+        await this.osc.setOscReceivingAddress(this.oscReceivingHost, this.oscReceivingPort);
       });
   }
 
@@ -136,7 +118,6 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
       .pipe(
         takeUntil(this.destroy$),
         distinctUntilChanged(),
-        switchMap((value) => combineLatest([of(value)]).pipe(map(([value]) => value))),
         tap(() => {
           this.oscSendingHostStatus = 'CHECKING';
           this.oscSendingHostError = undefined;
@@ -144,18 +125,8 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
         debounceTime(300)
       )
       .subscribe(async (host) => {
-        // Validate host
-        if (host === '' || !(isValidIPv6(host) || isValidIPv4(host) || isValidHostname(host))) {
-          this.oscSendingHostStatus = 'ERROR';
-          this.oscSendingHostError = 'invalidHost';
-          return;
-        }
-        // Save new host
         this.oscSendingHost = host;
-        this.oscSendingHostStatus = 'OK';
-        this.settingsService.updateSettings({
-          oscSendingHost: host,
-        });
+        await this.osc.setOscSendingAddress(this.oscSendingHost, this.oscSendingPort);
       });
   }
 
@@ -164,48 +135,16 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
       .pipe(
         takeUntil(this.destroy$),
         distinctUntilChanged(),
-        switchMap((value) =>
-          combineLatest([
-            this.settingsService.settings.pipe(
-              map((settings) => settings.oscSendingPort),
-              startWith(this.oscSendingPort),
-              distinctUntilChanged()
-            ),
-            of(value),
-          ]).pipe(map(([, value]) => value))
-        ),
         tap(() => {
           this.oscReceivingPortStatus = 'CHECKING';
           this.oscReceivingPortError = undefined;
         }),
         debounceTime(300)
       )
-      .subscribe(async (value) => {
-        // Parse port
-        const port = parseInt(value);
-        if (isNaN(port) || port > 65535 || port <= 0) {
-          this.oscReceivingPortStatus = 'ERROR';
-          this.oscReceivingPortError = 'invalidPort';
-          return;
-        }
-        // Validate port
-        if (port === this.oscSendingPort && this.oscReceivingHost === this.oscSendingHost) {
-          this.oscReceivingPortStatus = 'ERROR';
-          this.oscReceivingPortError = 'samePort';
-          return;
-        }
-        // Try to bind
-        if (!(await this.osc.init_receiver(this.oscReceivingHost, port))) {
-          this.oscReceivingPortStatus = 'ERROR';
-          this.oscReceivingPortError = 'bindFailed';
-          return;
-        }
-        // Save new port
-        this.oscReceivingPort = port;
-        this.oscReceivingPortStatus = 'OK';
-        this.settingsService.updateSettings({
-          oscReceivingPort: port,
-        });
+      .subscribe(async (port) => {
+        this.oscReceivingPort = parseInt(port);
+        if (isNaN(this.oscReceivingPort)) this.oscReceivingPort = 0;
+        await this.osc.setOscReceivingAddress(this.oscReceivingHost, this.oscReceivingPort);
       });
   }
 
@@ -214,43 +153,36 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
       .pipe(
         takeUntil(this.destroy$),
         distinctUntilChanged(),
-        switchMap((value) =>
-          combineLatest([
-            this.settingsService.settings.pipe(
-              map((settings) => settings.oscReceivingPort),
-              startWith(this.oscReceivingPort),
-              distinctUntilChanged()
-            ),
-            of(value),
-          ]).pipe(map(([, value]) => value))
-        ),
         tap(() => {
           this.oscSendingPortStatus = 'CHECKING';
           this.oscSendingPortError = undefined;
         }),
         debounceTime(300)
       )
-      .subscribe((value) => {
-        // Parse port
-        const port = parseInt(value);
-        if (isNaN(port) || port > 65535 || port <= 0) {
-          this.oscSendingPortStatus = 'ERROR';
-          this.oscSendingPortError = 'invalidPort';
-          return;
-        }
-        // Validate port
-        if (port === this.oscReceivingPort && this.oscReceivingHost === this.oscSendingHost) {
-          this.oscSendingPortStatus = 'ERROR';
-          this.oscSendingPortError = 'samePort';
-          return;
-        }
-        // Save new port
-        this.oscSendingPort = port;
-        this.oscSendingPortStatus = 'OK';
-        this.settingsService.updateSettings({
-          oscSendingPort: port,
-        });
+      .subscribe(async (port) => {
+        this.oscSendingPort = parseInt(port);
+        if (isNaN(this.oscSendingPort)) this.oscSendingPort = 0;
+        await this.osc.setOscSendingAddress(this.oscSendingHost, this.oscSendingPort);
       });
+  }
+
+  processValidation(validation: OscAddressValidation) {
+    this.oscReceivingHostError = validation.oscReceivingHost?.length
+      ? validation.oscReceivingHost[0]
+      : undefined;
+    this.oscReceivingHostStatus = validation.oscReceivingHost?.length ? 'ERROR' : 'OK';
+    this.oscReceivingPortError = validation.oscReceivingPort?.length
+      ? validation.oscReceivingPort[0]
+      : undefined;
+    this.oscReceivingPortStatus = validation.oscReceivingPort?.length ? 'ERROR' : 'OK';
+    this.oscSendingHostError = validation.oscSendingHost?.length
+      ? validation.oscSendingHost[0]
+      : undefined;
+    this.oscSendingHostStatus = validation.oscSendingHost?.length ? 'ERROR' : 'OK';
+    this.oscSendingPortError = validation.oscSendingPort?.length
+      ? validation.oscSendingPort[0]
+      : undefined;
+    this.oscSendingPortStatus = validation.oscSendingPort?.length ? 'ERROR' : 'OK';
   }
 
   login() {
@@ -266,5 +198,24 @@ export class SettingsVRChatTabComponent extends SettingsTabComponent implements 
     this.oscSendingPortChange.next(APP_SETTINGS_DEFAULT.oscSendingPort + '');
     this.oscReceivingHostChange.next(APP_SETTINGS_DEFAULT.oscReceivingHost);
     this.oscReceivingPortChange.next(APP_SETTINGS_DEFAULT.oscReceivingPort + '');
+  }
+
+  toggleOSCExpressionMenu() {
+    this.settingsService.updateSettings({
+      oscEnableExpressionMenu: !this.appSettings.oscEnableExpressionMenu,
+    });
+  }
+
+  toggleOSCExternalControl() {
+    this.settingsService.updateSettings({
+      oscEnableExternalControl: !this.appSettings.oscEnableExternalControl,
+    });
+  }
+
+  disableAllOSCFeatures() {
+    this.settingsService.updateSettings({
+      oscEnableExpressionMenu: !this.someOSCFeaturesEnabled,
+      oscEnableExternalControl: !this.someOSCFeaturesEnabled,
+    });
   }
 }